added startup notes

+++ td3algorithm/td3.py
@@ -89,7 +89,8 @@ class TD3(object):
             # Select action according to policy and add clipped noise
             noise = torch.FloatTensor(u).data.normal_(0, policy_noise).to(self.device)
             noise = noise.clamp(-noise_clip, noise_clip)
-            next_action = (self.actor_target(next_state) + noise).clamp(-self.max_action, self.max_action)
+            # train requires double, update to use float for better GPU processing
+            next_action = (self.actor_target(next_state.double()) + noise).clamp(-self.max_action, self.max_action)
 
             # Compute the target Q value
             target_q1, target_q2 = self.critic_target(next_state, next_action)
